+++ src/main/java/org/visuals/legacy/animatium/config/bundle/entry/EntryType.java
@@ -1,0 +1,9 @@
+package org.visuals.legacy.animatium.config.bundle.entry;
+
+public enum EntryType {
+    BOOLEAN,
+    INT,
+    FLOAT,
+    ENUM,
+    COLOR
+}